Updated type

+++ inst/include/PollardRho.h
@@ -1,6 +1,8 @@
 #ifndef POLLARDRHO_R
 #define POLLARDRHO_R 1
 
+#include <stdint.h>
+
 /**
  * Get Prime Factorization
  * t: number to factorize
@@ -8,6 +10,6 @@
  *
  * Note: this is adapted from demo "factorize.c" file from gmplib
  */
-void getPrimefactors (double& t, std::vector<double>&  factors);
+void getPrimefactors (int64_t& t, std::vector<int64_t>&  factors);
 
 #endif
